# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Discovery-маркетплейс специалистов: каталог, поиск, лиды, отзывы.",
        "title": "marketpclce API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/search/summarize": {
            "post": {
                "description": "По текстовому запросу/фильтрам поиска возвращает подобранную LLM-выдачу.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "llm"
                ],
                "summary": "LLM-подбор «топ-N специалистов»",
                "parameters": [
                    {
                        "description": "поисковые параметры",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_summarize.request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_summarize.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_summarize.errorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_summarize.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_summarize.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_summarize.Pick": {
            "type": "object",
            "properties": {
                "profile": {
                    "$ref": "#/definitions/marketpclce_internal_search.IndexDoc"
                },
                "rank": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_summarize.Result": {
            "type": "object",
            "properties": {
                "broadened": {
                    "type": "boolean"
                },
                "cached": {
                    "type": "boolean"
                },
                "picks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_summarize.Pick"
                    }
                },
                "summary": {
                    "type": "string"
                },
                "target_category": {
                    "type": "string"
                },
                "total_in_category": {
                    "type": "integer"
                },
                "usage": {
                    "$ref": "#/definitions/internal_summarize.Usage"
                }
            }
        },
        "internal_summarize.Usage": {
            "type": "object",
            "properties": {
                "cache_creation_input_tokens": {
                    "type": "integer"
                },
                "cache_read_input_tokens": {
                    "type": "integer"
                },
                "input_tokens": {
                    "type": "integer"
                },
                "output_tokens": {
                    "type": "integer"
                }
            }
        },
        "internal_summarize.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_summarize.request": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "city": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "q": {
                    "type": "string"
                },
                "rate_max": {
                    "type": "integer"
                },
                "rate_min": {
                    "type": "integer"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "target_category": {
                    "type": "string"
                }
            }
        },
        "marketpclce_internal_search.IndexDoc": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "city": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "is_published": {
                    "type": "boolean"
                },
                "primary_category": {
                    "type": "string"
                },
                "rate_max": {
                    "type": "integer"
                },
                "rate_min": {
                    "type": "integer"
                },
                "rating_avg": {
                    "type": "number"
                },
                "reviews_count": {
                    "type": "integer"
                },
                "skill_slugs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "skill_titles": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}